+++ examples/data-sources/sdwan_dns_security_policy/data-source.tf
@@ -1,4 +1,4 @@
-data "sdwan_dns_security_dns_security_policy" "example" {
+data "sdwan_dns_security_policy" "example" {
   id                 = "f6b2c44c-693c-4763-b010-895aa3d236bd"
   feature_profile_id = "f6dd22c8-0b4f-496c-9a0b-6813d1f8b8ac"
 }