+++ examples/resources/sdwan_dns_security_dns_security_policy/import.sh
@@ -1,2 +1,0 @@
-# Expected import identifier with the format: "dns_security_dns_security_policy_id,feature_profile_id"
-terraform import sdwan_dns_security_dns_security_policy.example "f6b2c44c-693c-4763-b010-895aa3d236bd,f6dd22c8-0b4f-496c-9a0b-6813d1f8b8ac"