+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Refresh devices of a configuration group before deletion, [link](https://github.com/CiscoDevNet/terraform-provider-sdwan/issues/391)
 - BREAKING CHANGE: Convert `feature_profiles` attribute of `sdwan_configuration_group` to set of strings
 - Add `sdwan_tag` resource and data source
-- Adds `sdwan_dns_security_dns_security_policy` resource and data source
+- Adds `sdwan_dns_security_policy` resource and data source
 
 ## 0.5.3